+++ src/onegov.ballot/HISTORY.rst
@@ -1,5 +1,7 @@
 Changelog
 ---------
+1.10.4 (2018-01-16)
+~~~~~~~~~~~~~~~~~~~
 
 - Adds a helper for getting the title translations.
   [msom]